+++ src/bundleReaders.h
@@ -1,5 +1,5 @@
-#ifndef _FIBERREADERS_H
-#define _FIBERREADERS_H
+#ifndef _BUNDLEREADERS_H
+#define _BUNDLEREADERS_H
 
 #include <cpp11.hpp>
 #include <string>
@@ -14,5 +14,10 @@ cpp11::writable::list ReadVTK(const std::string &inputTracts);
 cpp11::writable::list ReadVTP(const std::string &inputTracts);
 [[cpp11::register]]
 cpp11::writable::list ReadFDS(const std::string &inputTracts);
+[[cpp11::register]]
+cpp11::writable::list ReadTRK(const std::string &inputFile, int n_scalars,
+                              int n_properties, int n_count, bool little_endian,
+                              cpp11::strings scalar_names,
+                              cpp11::strings property_names);
 
-#endif /* _FIBERREADERS_H */
+#endif /* _BUNDLEREADERS_H */