+++ src/bundleWriters.cpp
@@ -1,4 +1,4 @@
-#include "fascicleWriters.h"
+#include "bundleWriters.h"
 #include <vtkDoubleArray.h>
 #include <vtkPointData.h>
 #include <vtkPolyDataWriter.h>